src/data_entry/data_entry_requirements.txt
streamlit>=1.31.0
supabase>=2.3.0
python-dotenv>=1.0.0
